Trait pallet_vanchor::pallet::Config

pub trait Config<I: 'static = ()>: Config + Config<I> {

    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type PalletId: Get<PalletId>;
    type LinkableTree: LinkableTreeInterface<LinkableTreeConfigration<Self, I>> + LinkableTreeInspector<LinkableTreeConfigration<Self, I>>;
    type KeyStorage: KeyStorageInterface<Self::AccountId>;
    type ProposalNonce: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen + From<Self::LeafIndex> + Into<Self::LeafIndex>;
    type VAnchorVerifier: VAnchorVerifierModule;
    type EthereumHasher: InstanceHasher;
    type IntoField: IntoPrimeField<AmountOf<Self, I>>;
    type Currency: MultiCurrencyExtended<Self::AccountId>;
    type PostDepositHook: PostDepositHook<Self, I>;
    type MaxExtAmount: Get<BalanceOf<Self, I>>;
    type MaxFee: Get<BalanceOf<Self, I>>;
    type MaxCurrencyId: Get<CurrencyIdOf<Self, I>>;
    type TokenWrapper: TokenWrapperInterface<Self::AccountId, CurrencyIdOf<Self, I>, BalanceOf<Self, I>, Self::ProposalNonce>;
    type NativeCurrencyId: Get<CurrencyIdOf<Self, I>>;
    type WeightInfo: WeightInfo;

}

The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type PalletId: Get<PalletId>

type LinkableTree: LinkableTreeInterface<LinkableTreeConfigration<Self, I>> + LinkableTreeInspector<LinkableTreeConfigration<Self, I>>

The tree type

type KeyStorage: KeyStorageInterface<Self::AccountId>

The key storage type

type ProposalNonce: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen + From<Self::LeafIndex> + Into<Self::LeafIndex>

Proposal nonce type

type VAnchorVerifier: VAnchorVerifierModule

The verifier

type EthereumHasher: InstanceHasher

The ethereum hash function for hashing external data (to match Solidity protocol)

type IntoField: IntoPrimeField<AmountOf<Self, I>>

A trait to map amount elements into a prime field.

type Currency: MultiCurrencyExtended<Self::AccountId>

Currency type for taking deposits

type PostDepositHook: PostDepositHook<Self, I>

An arbitrary execution function to execute after deposits/insertions are made

type MaxExtAmount: Get<BalanceOf<Self, I>>

Max external amount

type MaxFee: Get<BalanceOf<Self, I>>

Max fee amount

type MaxCurrencyId: Get<CurrencyIdOf<Self, I>>

Max currency ID value for signaling a strict transact without unwrapping

type TokenWrapper: TokenWrapperInterface<Self::AccountId, CurrencyIdOf<Self, I>, BalanceOf<Self, I>, Self::ProposalNonce>

TokenWrapper Interface

type NativeCurrencyId: Get<CurrencyIdOf<Self, I>>

Native currency id

type WeightInfo: WeightInfo

WeightInfo for pallet